pyproject.toml
[project]
name = "recallai"
version = "0.1.0"
description = "Recall AI - An AI-powered tool for information retrieval and management."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.1",
    "fastapi>=0.125.0",
    "langchain-text-splitters>=0.3.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pypdf>=6.4.2",
    "pyyaml>=6.0",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.38.0",
]
